requirements.txt
boto3==1.42.30
botocore==1.42.30
certifi==2026.1.4
charset-normalizer==3.4.4
colorama==0.4.6
fsspec==2026.1.0
idna==3.11
jmespath==1.0.1
numpy==2.4.1
pandas==2.3.3
pyarrow==23.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
s3fs==0.4.2
s3transfer==0.16.0
six==1.17.0
tqdm==4.67.1
tzdata==2025.3
urllib3==2.6.3
